Update with Google Login
Taylor made it so that Google+ Authentification now works. This update is incomplete for all the necessary things we need but I'm going to bed and wanted to upload to github. A Create Account page was made with no real functionality. The settings page was updated to Account Settings and features a functionless Change Password button and a logout button that I'm still not convinced logs the user out, but does kick them to the login page. We still need to add email and password login functionality, and a way to change that password. And I guess a way to check if they are logged in with google or not since we can't really change that password. Create a Card page now has a blank and an alert stating we want dynamic card creation in the future